Add osd-devices and osd-journal block storage
Add the "osd-devices" block-type storage, with
minimum of 0 and no maximum. Volumes assigned to
a unit will be added as OSDs.

Also, add the "osd-journal" block-type storage,
with minimum of 0 and maximum of 1. The osd-journal
storage, if supplied, will be used in favour of
osd-journal configuration if supplied.

For now, handling of osd-journal is static, just
as the configuration method was before. Removing
or adding the journal after deployment is not
currently supported. This is possible with further
changes, but requires stopping Ceph and migrating
the existing journal, and so out of scope here.