Add multiple pool support
This patch adds support for radosgw to be configured to talk to
a pool other than the default.  By setting the pool-prefix
the patch will create those pools after being related to the ceph
cluster.  This code follows the general format laid out on ceph'
s wiki for federated rados gateways.

Change-Id: If5e873647be136cd374f4bc8755a05f26a7399dc