Revert "Use libblkid instead of CLI"
This reverts commit 28740a3a1b8897073eaae64bdbcc3948b2512fe7
which switched to using libblkid instead of invoking the blkid
command on the commandline.

The replacement code introduced as part of that commit does not
actually fix the bug that it was intended to fix and only adds
unnecessary complexity to the code. The bug is being tagged
by this commit message to help identify the full set of fixes
for the actual bug.

Change-Id: I21b32197c8db4a3c667c986180cdd213fced2c26
Related-Bug: #1660240