Check existence of ceph binaries in add-storage hook
Add a check to make sure that the ceph package is installed in
the add-storage shim. The add-storage hook is run prior to the
install hook in order to provide storage for the install phase
of the charm in cases where it is needed. The ceph-osd charm
converts Juju storage into OSDs and doesn't need the storage
for the install hook, so just skip the hook if its run early.

Change-Id: I7b7518f52d0b5ad947b0809af8ad67d342211779
Closes-Bug: #1675186