keystone: PKI token format removal
As of Pike, the OpenStack charms no longer generate the certificates
and CA used to sign token revocation lists as this is associated with
the PKI token format, which has been removed from OpenStack in
favor of UUID or Fernet formats.

Soft-fail on cert retrieval if an InternalServerError is thrown; this
is most likely due to the fact that the keystone WSGI server cannot
find the relevant files on the underlying filesystem.

Change-Id: Ib592e7e47e10bed2d59c9136a3267f9c7ce8da83
Closes-Bug: 1709189