handle wal and db similarly to filestore journal
ceph-disk uses the same logic for partitioning if wal and db are put on
devices separate from the data device: it will use sgdisk to find the
next available partition to put wal or db to. This is no different from
the filestore partitioning behavior so we can just reuse this logic here

fs-type option is filestore-specific so let's filter it out if bluestore
is used. osd_format is a file system format for FileStore so this
becomes irrelevant.

also, rename least_used_journal finding function to a more generic name
to handle "utility" devices (journals, WALs, db-devices etc.)

Partial-Bug: #1710474
Change-Id: I483ee9dae4ce69c71ae06359d0fb96aaa1c56cbc