Ensure initial apparmor mode set
Due to changes in the hookenv.config charmhelper, the value of
aa-profile-mode does not change between the install and config-changed
hooks.  This results in the ceph-osd apparmor profile always
being enabled by default (rather than being disabled).

Ensure that an apparmor enforcement mode is correctly set whenever
a new profile is installed - this could either be on first install,
or if a new profile is added to the charm.

Change-Id: I131c9a871ad970b58fa6f41575c240081f653a21