Take control of when `ceph-create-keys` is run
For Ceph versions prior to Luminous the `ceph-create-keys`
is run by the init system in conjunction with starting the
ceph-mon service.

At some point during the Hammer lifetime a hard coded timeout
of 600 seconds was added:
https://github.com/ceph/ceph/commit/8bab74537f5891f2b97f0fb6722e2e00416b723a

In many circumstances this is not enough.  With this commit
we take control of running this command.

__NOTE__ it is important that charms deploying the ceph-mon
disable running of `ceph-create-keys` from the init system.

From Luminous and onwards we can use the newly added --timeout
flag.  For previous versions we rely on retry_on_exception
decorator.

Change-Id: I3d3c7298076730c423ca5cc059316619f415b885
Related-Bug: #1719436