Add support for RADOS gateway per unit cephx keys
If ceph-radosgw units present a key_name attribute, then generate
a cephx key using the key name and present back with the key_name
as the prefix for the key.

This switches radosgw units to having per-unit keys, rather than
sharing a global key (required as part of the pre-work to support
RADOS gateway federation).

Change-Id: I289b75a2935184817b424c5eceead16235c3f53b
Depends-On: I7ac4d23a91fa654b62afb62c389a8f9823ef2e05
Closes-Bug: 1808140