Fix nrpe ceph-osd status respecting permissions
The referenced bug (below) was caused because the nrpe check needed to
access the ceph owned directories, and as the nagios user, nrpe can't.
This change splits the check into a 'collect' phase that runs as root
via a cronjob each minute and writes a file to the tmp directory, and a
nrpe check phase that then reads that file and reports back to nagios.

The 'check' part deletes the 'collect' file, so that fresh information
is available for each nrpe check.  The cron task runs every minute (as
is lightweight), so the nrpe checks should not be sheduled more
frequently than 1 minute.

Change-Id: I4f4594a479eed47cc66643d0c6acece491ae854d
Closes-Bug: #1810749