Add standalone udevadm_settle function
Add a standalone udevadm_settle that can be called alone. This is
to enable it to be called by the ceph-osd charm. I think the
subprocess calls should be switched to check_call in the longer
term but that may be a disruptive change if udevadm does not
always return 0 so I propose to leave them as they are for the
moment.

Change-Id: Iec5932a4d819ad87e54c2af391abe1befe84f164
Partial-Bug: #1812925