Ensure we populate osd-devices with existing devices
If an older version of ceph-osd is deployed and then
upgraded to a version that keeps track of bootstrapped
OSDs, then the list of osd-devices never gets updated
with the pre-existing devices.

This change allows us to add existing, mounted Ceph
OSDs to the osd-devices entry in the local KV storage.

Change-Id: I940b108d914b39b55013a4617c3d17ff7122df60
Closes-Bug: #1814597