Switch auth order for s3 authentication
When deploying the RGW in multi-site configurations, communication
between sites is authenticated using S3 credentials managed within
RGW.  In the event that keystone authentication is in use this
generates a large number of s3 authentication attempts to keystone
which will always fail.

Switch the default order to check local auth first and then fallback
to external.

Change-Id: I7bfc016baf99188ba5a36f663145eeff465d25e8