Fix spurious nagios alerts for radosgw service.
Currently, when the charm tears down the default radosgw daemon in
order to make way for per host daemons, it does not remove the nrpe
check for the daemon. This PR fixes the issue.

It also closes a gap where alerts for the per host daemons are not
setup until a hook that happens to call update_nrpe_checks as a
side-effect is run.

Change-Id: I7621b9671b010a77bb3e94bdd1e80f45274c73e5
Closes-Bug: #1825843