Ensure that the mon relation completes
This change adds the fsid to the mon context
to ensure that the charm knows when the relation
is complete. Without an explicit key that only
comes from the relation, the ceph-radosgw charm
will happily go "ready" before that relation
has completed, leading to a confusing state
where the charm goes "blocked" because the
radosgw service is not running.

Change-Id: I4bfdabbd36400701debfb7a39a9c40701fc8b5ee