Add focal-ussuri and bionic-ussuri bundle
This patch adds a focal-ussuri and bionic-ussuri bundles to the tests
for the charm.

Also removes 'admin-token' from the tests, as the keystone charm no
longer accepts the option.

Also, for focal, btrfs-tools has been removed and replaced by
btrfs-progs; thus these are installed on focal instead.  See linked Bug.

Closes-Bug: #1871712
Change-Id: Ifc79beebb981571cebc38be6cb6c2748d22816c9