Add focal and ussuri bundles to the charm
This patch updates the bundles to include up to focal-ussuri.
The focal-ussuri bundle is in the dev bundles as it can't pass at the
moment due to LP: #1865754.

The bionic-ussuri bundle is in the dev bundles (i.e. not gate) as it
fails due to LP: #1892201

Also deal with the related bug where cinder-ceph requires the relation
with a nova-compute unit.

Related-Bug: #1881246
Related-Bug: #1865754
Related-Bug: #1892201
Change-Id: I0a6f1de82ecc601509822277d657485e08dc893d