Trigger udev rescan if pv_dev disappears
Workaround for kernel by in Ubuntu 20.04 LTS.

When using by-dname device paths with MAAS and bcache, the pvcreate
operation results in the by-dname entry for the block device being
deleted.  The subsequent vgcreate then fails as the path cannot
be found.

Trigger a rescan of block devices if the pv_dev path does not
exists after the pvcreate operation.

Change-Id: Ifb16c47ae5ff316cbcfc3798de3446a3774fa012
Related-Bug: 1878752