Remove chrony if inside a container
When running ceph-mon in containers, best practice is
to have chrony/ntp configured and installed on the bare
metal and then have the container trust the system
clock, as the container should not manage the system
clock.

The chrony package get installed automatically as
part of the dependencies of other packages, which
gets removed in this change.

ceph-mon charm will use this change here to remove
the chrony package.

Change-Id: Ie3c9c5899c1d46edd21c32868938d3290db321e7
Closes-Bug: #1852441