Enable passing of integer value for pool-set action
Enable passing of an integer to the pool-set function.  Due to how juju
appears to parse things on the command line, setting type to string
causes it to fail to accept '3', "3" or 3 as a string.  Only "'3'"
works.  However, if we remove the type from actions.yaml and do the
validation in the charm, any value can be passed.

Depends-On: I6081c23af61fd5e872982ff477b0a5cb27141d11

Change-Id: Idf3468d9ae28dafc09c86f08b7f8c6470a665b7a
Closes-Bug: #1838650