Add a progress watchdog for OSD upgrades
This patch add the concept of a watchdog to the upgrade_monitor so that
the charm can achieve two objectives of 1. Waiting for much longer, but
2. detecting whether the previous node has died / gone away.  This is
needed for 'large' OSDs where the time to upgrade a node may exceed the
current limit of 10 minutes, but also not to wait for 30 minutes on a
dead previous node.  The watchdog implements two timeouts and an
addition 'alive' key from the previous node to indicate that it is still
running.  Otherwise, functionality is identical.

Change-Id: Ia450e936c2096f092af3be5a369b7abaf5023b16
Related-Bug: #1762852