Change file owner so that check_ceph_osd nrpe service can work on CIS hardened environments
check_ceph_ods_services.py reads /var/lib/nagios file to report ceph
status back to nagios. This service runs as nagios user and the file
is owned by root. On CIS hardened servers the default mask is set to
027 making the permissions of the file 640 instead of 644.
This results in the service not being able to read the file and the
status reported to nagios is UNKNOWN even though ceph status is OK.

Closes-Bug: #1879667

Change-Id: Ib67b9a2b86a1c22658aeaf41f8e464072ab1828f