Align EC usage with best practice
When using erasure coding three pools are needed. One metadata
pool which is replicated, one default data pool which is replicated
and one erasure coded pool.

Add an additional ec-pool-weight option to allow the erasure coded
pool usage to be modelled within the deployment.

The consumer of the share will need to use xattrs to ensure that
the hierarchy of directories and files are stored in EC pool while
leaving the default replicated data pool for file backtraces.

Depends-On: Ib19e545fc676fa9e986f3fddcfb92cf99e778a3f
Depends-On: I48348ab2b6c8952c5e22008a074a60c1f35be952
Change-Id: I1af386d5e1d031f2f0387c7381af683f948172d9
Closes-Bug: #1895154