Support for iSCSI targets backed by an EC pool
Add support for creating an iscsi target backed by an EC pool.
The associated testing pr adds a test so that an EC backed
target is mounted in addition to a replication backed target.
This change also includes renaming the config option pool-name to
rbd-pool-name to make this charm more consistent with other charms
that relate to ceph.

Change-Id: I5772b4fc2415dd90029c3bde38bc57d490c9e910
func-test-pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/412